requirements.txt
addict==2.4.0
albumentations==1.3.1
aliyun-python-sdk-core==2.14.0
aliyun-python-sdk-kms==2.16.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
chumpy==0.70
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coverage==7.3.2
crcmod==1.7
cryptography==41.0.5
cycler==0.12.1
Cython==3.0.5
filelock==3.13.1
flake8==6.1.0
fonttools==4.44.0
fsspec==2023.10.0
idna==3.4
imageio==2.32.0
importlib-metadata==6.8.0
iniconfig==2.0.0
interrogate==1.5.0
isort==4.3.21
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
json-tricks==3.17.3
kiwisolver==1.4.5
lazy_loader==0.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe==0.7.0
mdurl==0.1.2
mmcv==2.1.0
mmdet==3.2.0
mmengine==0.9.1
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.28
ordered-set==4.1.0
oss2==2.17.0
packaging==23.2
pandas==2.1.2
parameterized==0.9.0
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
py==1.11.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.21
pycryptodome==3.19.0
pyflakes==3.1.0
Pygments==2.16.1
pyparsing==3.1.1
pytest==7.4.3
pytest-runner==6.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
requests==2.28.2
rich==13.4.2
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sympy==1.12
tabulate==0.9.0
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.9.26
toml==0.10.2
tomli==2.0.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.65.2
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
xdoctest==1.1.2
xtcocotools==1.14.3
yapf==0.40.2
zipp==3.17.0
